refactor: Update Movie and Person models
Updated the Movie and Person models to use mutable state instead of immutable interfaces.
This change simplifies the code and improves performance.

Additionally, the `PersonView` model now includes information about the movies the person has starred in.
The `MovieDetailView` model is now a data class with mutable properties, replacing the previous interface.

These changes make the models more flexible and easier to work with.